requirements.txt
alabaster==0.7.13
Babel==2.13.1
beautifulsoup4==4.12.2
bs4==0.0.1
builtwith==1.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
docutils==0.20.1
future==0.18.3
googlesearch-python==1.2.3
html5lib==1.1
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
Pillow==10.1.0
Pygments==2.16.1
PyPDF2==3.0.1
python-whois==0.8.0
requests==2.31.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
urllib3==2.0.7
webencodings==0.5.1
